refactor: 로직 분리 (데이터 관리-Page , UI-views)
