+++ src/pages/todo/index.tsx
@@ -1,25 +1,15 @@
-import { useCallback, useMemo, useState } from "react";
+import { useCallback } from "react";
 
-import TodoList from "@/components/organisms/todo-list/TodoList";
 import { useTodos } from "@/hooks/todo/useTodos";
 import { useUpdateTodo } from "@/hooks/todo/useUpdateTodo";
 import { cn } from "@/utils/cn";
+import Todos from "@/views/todo/todoPage/Todos";
 
 const HEADER_HEIGHT = 48;
-const FILTER_TYPES = ["All", "Done", "To do"] as const;
 
 export default function TodoPage() {
   const { data, isLoading, error } = useTodos();
   const toggleTodoMutation = useUpdateTodo();
-
-  const [filter, setFilter] = useState<(typeof FILTER_TYPES)[number]>("All");
-
-  const filteredTodos = useMemo(() => {
-    if (!data) return [];
-    return data.todos.filter((todo) =>
-      filter === "Done" ? todo.done : filter === "To do" ? !todo.done : true,
-    );
-  }, [data, filter]);
 
   const handleToggleTodo = useCallback(
     (todoId: number, isDone: boolean) => {
@@ -55,32 +45,7 @@ export default function TodoPage() {
         </button>
       </div>
 
-      <div className="mb-4 flex h-full flex-grow flex-col rounded-xl bg-white p-4 sm:mb-6 sm:p-6 md:max-w-[744px]">
-        <div>
-          {FILTER_TYPES.map((type) => (
-            <button
-              key={type}
-              onClick={() => setFilter(type)}
-              className={cn(
-                "mr-2 mb-4 rounded-full border border-slate-200 px-3 py-1 text-sm font-medium hover:shadow-sm",
-                filter === type && "border-blue-500 bg-blue-500 text-white",
-              )}
-            >
-              {type}
-            </button>
-          ))}
-        </div>
-
-        {filteredTodos?.length > 0 ? (
-          <TodoList data={filteredTodos} handleToggleTodo={handleToggleTodo} />
-        ) : (
-          <div className="flex flex-1 items-center justify-center text-sm font-normal text-slate-600">
-            {filter === "All" && "등록한 일이 없어요"}
-            {filter === "Done" && "다 한 일이 아직 없어요"}
-            {filter === "To do" && "해야할 일이 아직 없어요"}
-          </div>
-        )}
-      </div>
+      {data && <Todos todos={data.todos} onToggleTodo={handleToggleTodo} />}
     </div>
   );
 }
